+++ models/following.go
@@ -5,6 +5,6 @@ import "gorm.io/gorm"
 type Following struct {
 	gorm.Model
 
-	HostID   int64
-	FollowID int64
+	HostID   uint
+	FollowID uint
 }
